rework .ova generation
The .ova now contains two empty disks, a root disk and a data disk.
It also contains an installation ISO. On first boot (when the root
disk is empty) the ISO will boot and ask the user for permission to
install. The install script partitions and formats all disk (wiping
all existing data) and copies the root filesystem.

This also replaces the old first-boot scripts.